à fond sur les observables ..

+++ SimpleCRM.App/src/app/main/entity/list/entity-items-list-details/entity-items-list-details.component.ts
@@ -7,20 +7,31 @@ import {
   ViewChild,
   TemplateRef,
   Output,
-  EventEmitter
+  EventEmitter,
+  Input
 }                                                 from '@angular/core';
 import { FormGroup }                              from '@angular/forms';
 import { MatDialogRef, MatDialog, MatSort, Sort } from '@angular/material';
+import { Store, select }                          from '@ngrx/store';
 import { Subject, Observable }                    from 'rxjs';
-import { takeUntil }                              from 'rxjs/operators';
+import { takeUntil, tap, filter }                 from 'rxjs/operators';
 
 import { fuseAnimations }                         from '@fuse/animations';
 import { FuseConfirmDialogComponent }             from '@fuse/components/confirm-dialog/confirm-dialog.component';
 
-import { EntityService }                          from '../../entity.service';
-import { EntityItemsListItemFormDialogComponent } from '../entity-items-list-item-form/entity-items-list-item-form.component';
 import { IItem }                                  from 'app/models';
+import { EntityItemsListItemFormDialogComponent } from '../entity-items-list-item-form/entity-items-list-item-form.component';
+import { EntityService }                          from '../../entity.service';
+import { ElementsState, entitySelectors }         from '../../store';
 
+/**
+ * The main EntityItemsListDetailsComponent class
+ *
+ * @export
+ * @class EntityItemsListDetailsComponent
+ * @implements {OnInit}
+ * @implements {OnDestroy}
+ */
 @Component({
   selector: 'entity-items-list-details',
   templateUrl: './entity-items-list-details.component.html',
@@ -29,6 +40,14 @@ import { IItem }                                  from 'app/models';
   animations: fuseAnimations
 })
 export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
+
+  /**
+   * Sort options (orderby ==> column, direction)
+   *
+   * @type {Sort}
+   * @memberof EntityItemsListDetailsComponent
+   */
+  @Input() orderBy: Sort;
 
   /**
    * Event emitter to transmit sort changes to parent component
@@ -59,7 +78,6 @@ export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
   items: any;
   user: any;
   dataSource: FilesDataSource | null;
-  dataSource2: FilesDataSource2 | null;
   // displayedColumns = ['checkbox', 'avatar', 'id', 'name', 'active', 'created', 'buttons'];
   displayedColumns = ['checkbox', 'avatar', 'id', 'active', 'created', 'buttons'];
   selecteditems: any[];
@@ -68,7 +86,7 @@ export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
   confirmDialogRef: MatDialogRef<FuseConfirmDialogComponent>;
 
   // private
-  private _unsubscribeAll: Subject<any>;
+  private _unsubscribeAll: Subject<any> = new Subject();
 
   /**
    * Constructor
@@ -78,12 +96,9 @@ export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
    */
   constructor(
     private _entityService: EntityService,
-    public _matDialog: MatDialog
-  ) {
-
-    // set the private defaults
-    this._unsubscribeAll = new Subject();
-  }
+    public _matDialog: MatDialog,
+    private _store: Store<ElementsState>
+  ) {}
 
   // -----------------------------------------------------------------------------------------------------
   // @ Lifecycle hooks
@@ -95,79 +110,59 @@ export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
   ngOnInit(): void {
 
     // instantiate data source
-    this.dataSource = new FilesDataSource(this._entityService);
-    this.dataSource2 = new FilesDataSource2(this._entityService);
+    this.dataSource = new FilesDataSource(this._store);
 
-    // listen to onItemsChanged
-    this._entityService.onItemsChanged2
-
+    // get current displayed items from store
+    this._store.select(entitySelectors.getCurrentDisplayedItems).pipe(
       // attach unsubscriber
-      .pipe(takeUntil(this._unsubscribeAll))
-
-      // subscribe
-      .subscribe(items => {
-
-        // save items
-        this.items = items;
-
-        // uncheck all checkboxes
-        this.checkboxes = {};
-        items.map(item => this.checkboxes[item.id] = false);
-      });
+      takeUntil(this._unsubscribeAll),
+      // if are defined
+      filter(items => items !== undefined),
+      // uncheck all checkboxes
+      tap(items => items.map(item => (this.checkboxes = {}) && (this.checkboxes[item.id] = false)))
+    ).subscribe();
 
     // listen to onSelectedItemsChanged
     this._entityService.onSelectedItemsChanged
-
       // attach unsubscriber
       .pipe(takeUntil(this._unsubscribeAll))
-
       // subscribe
       .subscribe(selecteditems => {
-
         // for each checkbox
         for (const id in this.checkboxes) {
-
           // assert id
           if (!this.checkboxes.hasOwnProperty(id)) continue;
-
           // check checkbox if in selected items
           this.checkboxes[id] = selecteditems.includes(id);
         }
-
         // save selected items
         this.selecteditems = selecteditems;
       });
 
-    // listen to onUserDataChanged
-    this._entityService.onUserDataChanged
-
-      // attach unsubscriber
-      .pipe(takeUntil(this._unsubscribeAll))
-
-      // subscribe to save user
-      .subscribe(user => this.user = user);
+    // // listen to onUserDataChanged
+    // this._entityService.onUserDataChanged
+    //   // attach unsubscriber
+    //   .pipe(takeUntil(this._unsubscribeAll))
+    //   // subscribe to save user
+    //   .subscribe(user => { console.log(user); this.user = user; });
 
     // listen to onFilterChanged
-    this._entityService.onFilterChanged
-
+    this._entityService.onFilterChanged2
       // attach unsubscriber
       .pipe(takeUntil(this._unsubscribeAll))
-
       // subscribe to deselect all items of signal
       // .subscribe(() => this._entityService.deselectItems());
       .subscribe(this._entityService.deselectItems.bind(this._entityService));
 
     // listen to sort changes of the mat-table
     this.sort.sortChange
-
       // attach unsubscriber
       .takeUntil(this._unsubscribeAll)
-
       // emit to parent component
       .do(this.changeSort)
-
       // subscribe
       .subscribe();
+
   }
 
   /**
@@ -247,20 +242,26 @@ export class EntityItemsListDetailsComponent implements OnInit, OnDestroy {
   /**
    * Delete item
    *
-   * @param {*} item item to delete
+   * @param {any} item item to delete
    * @memberof EntityItemsListDetailsComponent
    */
-  deleteItem(item): void {
+  deleteItem(item: any): void {
+
+    // open closeable dialog and save reference
     this.confirmDialogRef = this._matDialog.open(FuseConfirmDialogComponent, {
       disableClose: false
     });
 
+    // set confirm message for delete
     this.confirmDialogRef.componentInstance.confirmMessage = 'Are you sure you want to delete?';
 
+    // on close dialog
     this.confirmDialogRef.afterClosed().subscribe(result => {
-      if (result) {
-        this._entityService.deleteItem(item);
-      }
+
+      // delete item if result
+      if (result) this._entityService.deleteItem(item);
+
+      // remove reference to dialog
       this.confirmDialogRef = null;
     });
 
@@ -308,11 +309,11 @@ export class FilesDataSource extends DataSource<IItem> {
   /**
    * Constructor
    *
-   * @param {EntityService} _entityService main service for entities and entity items
+   * @param {Store<ElementsState>} _store store for elements (entities, items)
    * @memberof FilesDataSource
    */
   constructor(
-    private _entityService: EntityService
+    private _store: Store<ElementsState>
   ) {
 
     // invoke base constructor
@@ -328,8 +329,8 @@ export class FilesDataSource extends DataSource<IItem> {
    */
   connect(): Observable<IItem[]> {
 
-    // bind to onItemsChanged subject of our entities service
-    return this._entityService.onItemsChanged;
+    // bind to an elements state selector
+    return this._store.pipe(select(entitySelectors.getCurrentDisplayedItems));
 
   }
 
@@ -339,44 +340,5 @@ export class FilesDataSource extends DataSource<IItem> {
    * @memberof FilesDataSource
    */
   disconnect(): void { }
-
-}
-
-export class FilesDataSource2 extends DataSource<any> {
-
-  /**
-   * Constructor
-   *
-   * @param {EntityService} _entityService
-   * @memberof FilesDataSource2
-   */
-  constructor(
-    private _entityService: EntityService
-  ) {
-
-    // invoke base constructor
-    super();
-
-  }
-
-  /**
-   * Connect function called by the table to retrieve one stream containing the data to render.
-   *
-   * @memberof FilesDataSource2
-   * @returns {Observable<any[]>}
-   */
-  connect(): Observable<any[]> {
-
-    // bind to onItemsChanged2 subject of our entities service
-    return this._entityService.onItemsChanged2;
-  }
-
-  /**
-   *
-   * Disconnect implementation
-   *
-   * @memberof FilesDataSource2
-   */
-  disconnect(): void {}
 
 }
